kclient: fix cap_snap flush to reference correct snap
The client was passing the snapid the data was to fall into
instead of the snap it _follows_ (which is now the mds tends
to look at things).  Use snapid-1 instead.  This fixes the
problem of files being written while the snapshot was taken
appearing as 0 bytes in the snapshot.