kclient: add dput to getattr for do_lookup splice case
If do_lookup returns a dentry, it has a new ref on it.  Make sure we
dput() that, since we don't actually need it.  (In contrast, if there is no
splice, do_lookup does not dget, as the caller has a ref on the dentry it
provided in the first place.)