kclient: flush/release all caps back to mds before closing a session
This allows us to handle races that can occur if an inode moves between
MDSs.  It also ensures that _all_ dirty cap data is safely flushed before
we throw it out.

To make this work, we drop the PIN reference on the subdirectory that
we mounted.  Since VFS doesn't tell us when each vfsmount goes away,
we can't remove them as we go.  And we're more or less screwed anyway
if the directory gets renamed (currently, at least), so there's not
much point in having this reference in the first place.