mds: fix filelock loner modes
Rewrote Filelock handling of "loner" mode, which gives a single client access
to do cached/bufferd read/write on a file.  To make this sane, CInode::loner_cap
identifies the distinguished client (if any), and the FileLock caps methods
provides issued/allowed/wanted values for loner and other clients.  A new
lock state is added to track revocation of RDCACHE from lock->loner on the
non-loner clients.

Overall, the new code is much cleaner and should be less susceptible to
odd corner cases...