osd: repeer osd if prior set may be affected
Previously we only repeered if the active set changed.  However, changes
in the up/down state of the prior set (or prior set candidates) or the
primary osd's up_thru can also affect the prior set and peering.

This fixes the problem where PGs get stuck in a "crashed" state without
moving to "crashed+replay".  We sit and wait for info from a peer who
we thought was up but is now down, or vice-versa.