mds: unqueue recovery on purging inodes
If an inode is queued for file size recovery when it is purged,
unqueue it.  This catches the log replay case where client
reconnect queues up the inode.

Also, in eval_stray, skip inodes that are queued.  This should
avoid a recovery running concurrently with the purge (which could
be problematic, as it would carry a pointer to *in).