mds: make open_remote_ino terminate if the anchortrace refers to a non-existent ino
Since anchor lookup is racy, we may have to do multiple lookups
(in the case of a concurrent anchor table update).  If we don't
find the ino in the anchor, remember the anchor version when we
try again.  If we fail again at the same point, and the anchor
has not changed, fail.

Create an open_remote_dentry helper that does this.  If we fail,
set the CDentry::STATE_BADREMOTEINO state bit.