mds: fix up completed_request handling during journal replay
The completed_requests is handled separately from the session table
itself, in that we may add completed requests to the table even when
we may have loaded newer info.  But the handling was a bit wrong.
We make sure we only add completed requests if the session is already
open... and remove the unnecessary trim (if the sessionmap is newer, the
session is already closed, and thus we have no request info).