osd: use last_clean_interval in build_prior logic
We now mark a PG crashed if any of the OSDs during a given interval
is not either still alive or cleanly shut down during the interval.  If
those two conditions are not yet, it may have crashed.

It isn't a perfect set of criteria, since last_clean_interval is only
clean shutdowns of an OSD.  We could, for example, track another interval
generated via old osd_up_thru, but the conditions for that are different,
since that requires survival past the end of the interval, not a clean
shutdown during the interval.  This should capture the common case, though,
of a clean unmount.
to