mds: fix try_remove_unlinked_dn
Just because a dir is_complete() doesn't mean that we can safely
do away with a null dentry.  What matters is if the dentry existed
before the last commit.. and having everything in our cache isn't
relevant at all there.  If we are complete, a fetch() would be a
no-op, but since we can clear complete at any time, that's not
helpful.