kclient: use helper to clearly define which caps belond on rdcaps
Adjust list membership in ceph_add_cap, release, flush_ack, and
ceph_put_fmode paths.  Use a spinlock to protect list adjustments
because ceph_put_fmode does not have the session mutex.

Make __cap_is_valid() to check if we are on the rdcaps list or not.

Remove unneeded extra checks in rdcaps trimmer; replace with a
WARN.