kclient: implement permission iop
The permission iop verfies we have a valid uid/gid/mode by checking for
the AUTH cap.  Then calls the generic helper.

This is 100% consistent because we don't carry a cap ref over the
permission checkout.  OTOH, if we don't hold the cap and the mds doesn't
actually issue it, we never will anyway.  Maybe fix this eventually.