osd: break apart write stages, transactions
We break the write preparation into three stages.  First we run the ops
vector and build the op_t transaction.  If it is non-empty, we build a
clone_t transaction to run before it, and a local_t that updates the osd's
PG log and metadata.

Take care to preserve old exists, size, and version values before running
the ops vector as those are clobbered but need to be send to the replica
osds.