osd: track ObjectContext for cloned objects
This orders access to a newly cloned object.  This is really only important
when you have a racing clone creation and a clone read.  The read will
look at the head's context and expect the clone to be there, but we may
not have applied the write to disk yet.  So, we set up an obc for the
cloned object too (with the same mode as the head).