osd: ship transaction (not op) to replicas
This simplifies the code path on the replica.  It avoids duplicating
computation, which can be a win even for trivial ops when snaps are
involved.  For active objects, it will avoid the computation as well.

Currently parallel execution will still happen if you specify a flag.
However, an object method should probably also be able to specify ||
execution is better in cases where the operation is deterministic and
|| exec will mean less data over the network.

Disable unused vector<T*> template, since it confuses the
vector<const char *> encoder.