kclient: move max_size update logic into ceph_get_cap_refs; fix race
It was previously possible to check for max_size, see it was large
enough, unlock i_lock, and go into the cap wait loop.  If a racing
thread reduces max_size, we will spin forever.

Break the cycle by moving hte max_size logic into get_cap_refs.  If
we are waiting and need a larger max_size, set a flag, so that we
recheck (possibly calling ceph_check_caps) and wait again.