rados: include pool_snap_info in pg_pools
Include info about snaps global to each pg_pool in the pg_pool_t (and
thus OSDMap).  These snap contexts will be used when a snapc is not
explicitly specified by the client.

Note that the pool-wide snaps are (presently) mutually exclusive with
client-controlls snaps (as used by ceph) because a mixture of @snaps and
@removed_snaps doesn't clearly defined which snaps do not exist (they
share the namespace).