add Completion class
This should eventually replace Context.  The main difference is instead
of

   context->finish(r);
   delete context;

you do

   completion->complete(r);

and the completion should delete this as needed.  It's more flexible
because it doesn't require a new heap allocation for each caller
reference.  And 'delete this' is legal, so long as you do not access any
member fields after.