osd: you may, but need not, specify READ|WRITE flag in MOSDOp.
The OSD will implicitly set the bits based on your OSDOps or class method
calls.  The client may still find it useful to specify these expicitly
for it's own informational purposes.

Make sure the MOSDOpReply has bits set based on the _actual_ op performed.

Note that as things stand, this will confuse the Objecter, who relies on
these bits to choose read or modify reply paths and doesn't know a priori
what mode a method is.