osd: fix initialization of log.complete_to in PG::activate()
The complete_to should point to the next object to get, which
should be just PAST info.last_complete.  That is because we
can trim the log up to and including last_complete (because
that entry is recovered), and we don't want to invalidate
the iterator.

That is
    while (log.complete_to->version <= info.last_complete)
      log.complete_to++;

and in sub_op_push,

    while (...) {
      ...
      if (info.last_complete < log.complete_to->version)
	info.last_complete = log.complete_to->version;
      log.complete_to++;
    }