osd: force RMW ordering globally
We can't mix RMW and DELAYED in the same PG without screwing
up the ordering of writes, the pg log, and so forth.

So force RMW throughout.  This won't affect the mds log
appends because the client is constant.  It will slow down
concurrent writes to the same object by multiple clients, but
we don't have many (any?) of those yet.

This needs a real solution... :/