osd: make write mode per-PG
We can't do it per-object because the access mode determines the order
we append to the log, and that has to be sequential.  It has to be per-PG,
unless a whole ton of other stuff is reworked.

This lets us capture the best access mode at least on a per-pool basis,
instead of imposing a global default.