mon: new, improved monitor request/reply routing
We track which monitor requests are forwarded to the leader in the mon
Session, and clean them out when we send a reply back via MRoute.

Only forward once, and resend pending requests on election completion.
This may lead to harmless dups, but at least we know our request didn't
get dropped on the floor by another mon failing somewhere.

This changes the monitor protocol by expanding the PaxosServiceMessage
header.