osd: eval access mode before try_*
The problem is the try_* functions may change the mode, but a start_write
may not follow if the op doesn't execute (for any number of reasons).  So
the next time we come around, we may need to reset back to IDLE if
num_wr == 0.  Be careful about the wake flag.

This fixes osd op hangs (particularly after mds restart, and lots of
failed stats on objects during file size/mtime recovery).