mds: do not eval subtree root during rename_prepare
The eval may lead to a scatter_writebehind, which screws up the journal
ordering.  Since the inode is also locked, we will end up evaling it
later anyway, so don't worry about it.

I'm a bit unsure about the adjust_subtree_after_rename... that is during
the rename_apply stage (after the log entry committed), so i think it's
safe to do a scatter_writebehind here.. since we're not between a predirty
and submit_entry.