mon: stop trying to create localized pgs >= max osd|device
If lpg_num == lpgp_num, and the preferred osd no longer exists, there is
no guarantee that the parent and child pgs will land on the same osd.  This
happens when we are creating pgs for a new osd, they don't finish getting
created, and then we destroy that OSD again.

Care should be taken to avoid expanding the OSD count _and_ lpg_num
simultaneously...