mds: add NEW session state
We want to register new sessions in the map so that they can be found when
we replay ESession events in the log.  Otherwise a client that connects
to the mds early will have one Session* on the connection, and another in
the map, and reconnect will fail.

Make sure we remove NEW sessions from the map on connection reset, so that
things gets freed.

Also, take mds_lock in ms_verify_authorizer.