mds: expose projected ctime is client is last versionlock wrlocker
This fixes the problem where the client does a setattr on CTIME, but
doesn't actually see it.

Probably, though, the exposed ctime should be more precise: something like
the most recent projected value that the client is responsible, since we
may have multiple clients projecting pipelined values on the inode.