filestore: fix up commit_op_seq
Note that we now _only_ update commit_op_seq when doing a proper commit.
That means that any additional commit the fs does will be unknown to us. If
we have a journal, that's fine, we'll replay it.  If not, we're screwed.

So if there's no journal, we need to be using btrfs_snap mode, and roll
back to the last snap.