osd: wait for readable on primary after pulling an object
Otherwise our attempt to push it back out again fail.

Note that currently this disables the previous push-right-after-
pull behavior.  We should either optimize the common case (full,
non-snapped object), or at least requeue the object so it's at
the front of the recovery queue so the fs cache will be hot.