ceph: fix manpages so they are only installed once
While creating a spec file for CEPH, rpmbuild was complaining because make
install was copying the manpages in, and then copying them in again.  This is
because man_MANS and dist_man_MANS are supposed to be two seperate lists that do
not overlap.  So make install would install all the man pages in the man_MANS
list and the dist_man_MANS list.  This patch kills the dist_man_MANS thing to
keep this from happening.  This made rpmbuild happy, which makes me happy :).
Thanks,

Signed-off-by: Josef Bacik <josef@redhat.com>