objectstore: simpler transaction encoding
Just concatenate operations to a bufferlist as we go.  No
distinct decoding step is needed; we parse the transaction as it
is replayed/applied.  This avoids the old decoded intermediate
representation overhead.

Since we still decode the old version, that code is still there,
but not used for anything new.