mds: put forced open sessions in OPENING then OPEN
We use OPENING state to indicate sessions that are being
imported.  Fix get_or_add_open_session() to NOT set the session
state (except to STATE_NEW if new) so that the caller can do
the right thing.  Otherwise, the prepare_force_open_sessions()
can't tell if it just forced open a session (and needs it to be
OPENING) or if it was already open.  Subsequently cap migrations
weren't working if the client didn't already have a session
open.

There is still a bug: if the import aborts, we have an OPENING
session with no actual open client_session message queued.  Maybe
we should have a different state instead of OPENING... IMPORTING?