journal: set committing_seq in commit_start
We need to set committing_seq when we first block access to the
fs in commit_start(), NOT in commit_started(), as by that time
we may have queued all sorts of additional entries for the
journal.  The committing_seq is all about syncing up with the
live fs (via op_apply_start), not what's queued for the journal.