mds: accept stray reconnects
Hmm.  Ultimately this is the direction we want to go, so that,
for example, a client that is forcefully disconnected can at
least attempt a reconnect (even if all its caps go ESTALE).
There are undoubtably some issues that need to be dealt with to
keep the mds from choking on weird client data, but those things
need to be fixed regardless.