context: don't complete C_Gather until finisher is set
We often create a C_Gather, add a bunch of subs, and set a
finisher at the end.  Occasionally we create a sub and it gets
finished immediately, in which case we don't want the gather to
get finished and deleted already.  So don't 'complete' the gather
until it's by activated by setting the finisher.